003: say what the conversion ratchet is made of
`conversion warnings 1156` says how many are left and not whether any can
be fixed, which is the same defect the `goto` row had before it got its
breakdown.  So:

    conversion warnings                1156
      signedness, same width            651
      narrowing 32 -> 16                306
      narrowing 32 -> 8                 167
      ...
      a negative constant into an unsigned type    5

The rows add up to the ratchet by construction -- there are rows for
`-Wsign-compare` and `-Wuseless-cast` that are currently zero and absent.
The first version dropped two `useless-cast` messages on the floor and the
breakdown came to 1156 against a ratchet of 1158, which is the kind of
measurement §10 is about.

The distinction the row makes is the one that matters: 502 of these are
narrowings into fields the 1997 layout fixes at eight or sixteen bits, and
no declaration can fix those.  651 are signedness at the same width, which
is `resign.py`'s business and where a later round has something to work
with.

Five negative constants reaching unsigned types now say they mean it.
`fold[128] = (uint8_t)-1` is the fold table's odd codes counting down from
-1, which in a byte is 255 -- the same test the comparison casts pass, in
that the compiler naming the value is not the same as the reader knowing
which.  A sixth was itself useless and came straight back out.

Gate green.  Sweep: exit 0.

+++ 003/subs1.hpp
@@ -1658,9 +1658,9 @@ uint32_t __alt_init_tables(uint8_t *fold, int8_t *unfold)
   for ( i = 0; i < 0x3F; ++i )
   {
     ((uint8_t *)unfold)[4 * i + 2] = 2 * i + 1;
-    ((uint8_t *)unfold)[4 * i + 1] = -2 * i - 1;
+    ((uint8_t *)unfold)[4 * i + 1] = (uint8_t)(-2 * i - 1);
     ((uint8_t *)unfold)[4 * i + 4] = 2 * i + 2;
-    ((uint8_t *)unfold)[4 * i + 3] = -2 * i - 2;
+    ((uint8_t *)unfold)[4 * i + 3] = (uint8_t)(-2 * i - 2);
   }
   unfold[254] = 127;
   unfold[253] = -127;
@@ -1676,7 +1676,10 @@ uint32_t __alt_init_tables(uint8_t *fold, int8_t *unfold)
   // 22-line else are gone.  Same argument as the block above it.
   *fold = 0;
   lo = 1;
-  fold[128] = -1;
+  // The wrap is the point: the odd codes count down from -1, which in a byte
+  // is 255.  The cast says so, where the bare `-1` left the compiler to point
+  // out that the value changes -- which it does, deliberately.
+  fold[128] = (uint8_t)-1;
   // -E is 0, so the near-lossless fill that stood here never ran: the jump
   // over it was `if ( 1 ) goto LABEL_52`, and LABEL_52 had no other source.
   // The `lo < 128` test it jumped past went with it -- the only live
